# niftyNotaryAPI/tests/swagger.yml
...
openapi: "3.0.0"
info:
  version: 0.0.1
  title: niftyNotary API
  license:
    name: MIT
servers:
  - url: http://localhost:3600
paths:
...
  /notarised/getByTxId/{txId}:
    get:
      summary: Get notary object from db
      operationId: fetchTxByTxId #what is the operationId: for???
      tags:
        - notary
      parameters:
        - name: txId
          in: path
          required: true
          description: The tx id of the item to retrieve
          schema:
            type: string
          example: '0x1111111111111111111111111111111111'
      responses:
        '200':
          description: todo should send back the same details as post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notarised"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notarised:
      type: object
      properties:
        txStatus:
          type: string
        fileHash:
          type: string
        hashType:
          type: string
          enum: [Invalid, Exists, Sha256, Sha3, EthSig, Sha3PlusEthSig, Passphrase, Other] #this should probably be its own object
          example: 'Sha3'
        link:
          type: string
        signature:
          type: string
#        docType:
#          type: string
        txId:
          type: string
        chainId:
          type: string
          enum: [1]
        contractAddress:
          type: string
        timestamp:
          type: string
        id:
          type: string
        explorer:
          type: string
...